+++ types/index.ts
@@ -1,0 +1,4 @@
+export type SubjectColor = {
+  filterId: number;
+  color: string;
+};
